+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+### v1.2 (29.09.2023)
+ - Dictionary has now higher priority than DevelopmentLanguageId
+ - Dictionary is now alfabetically sorted using [System.Collections.Specialized.OrderedDictionary]
+ - Added [B]reak option to ExtendedStringConfirmation
+ - Added progress indicator; files without changes are not reported to the console
+ - Fixed bug when saving DeepL credentials; .deepl folder created if not exists
+ - Fixed bug when saving dictionary; .dictionary folder created if not exists
+ - Added .gitignore with /.deepl and /.dictionary folders
+ - Other minor changes
+
 ### v1.1 (06.09.2023)
  - Adding Manifest to give more information about the module
  - Importing depenency module Microsoft.Dynamics.Nav.Model.Tools directly